requirements.txt
torch
transformers
datasets
evaluate
textattack
scikit-learn
pandas
numpy
pyyaml
tensorflow-hub
